requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.11.3
psycopg2==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
whitenoise==5.0.1
wrapt==1.12.0
